requirements.txt
gymnasium==0.29.1
minigrid==2.3.1
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
tensorboard>=2.13.0
tqdm>=4.65.0
pillow>=10.0.0
imageio>=2.31.0